app/requirements.txt
fastapi
uvicorn
requests
python-docx
PyMuPDF
